+++ lonboard/_viz.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import json
 from random import shuffle
+from textwrap import dedent
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -30,7 +31,7 @@ from lonboard._geoarrow.parse_wkb import parse_wkb_table
 from lonboard._geoarrow.sanitize import remove_extension_classes
 from lonboard._layer import PathLayer, PolygonLayer, ScatterplotLayer
 from lonboard._map import Map
-from lonboard._utils import get_geometry_column_index
+from lonboard._utils import get_geometry_column_index, split_mixed_gdf
 from lonboard.basemap import CartoBasemap
 from lonboard.types.layer import (
     PathLayerKwargs,
@@ -40,6 +41,7 @@ from lonboard.types.layer import (
 from lonboard.types.map import MapKwargs
 
 if TYPE_CHECKING:
+    import duckdb
     import geopandas as gpd
 
     class GeoInterfaceProtocol(Protocol):
@@ -66,6 +68,7 @@ if TYPE_CHECKING:
         ArrowStreamExportable,
         GeoInterfaceProtocol,
         Dict[str, Any],
+        duckdb.DuckDBPyRelation,
     ]
     """A type definition for allowed data inputs to the `viz` function."""
 
@@ -95,6 +98,7 @@ def viz(
     path_kwargs: Optional[PathLayerKwargs] = None,
     polygon_kwargs: Optional[PolygonLayerKwargs] = None,
     map_kwargs: Optional[MapKwargs] = None,
+    con: Optional[duckdb.DuckDBPyConnection] = None,
 ) -> Map:
     """A high-level function to plot your data easily.
 
@@ -103,15 +107,66 @@ def viz(
 
     This function accepts a variety of geospatial inputs:
 
-    - geopandas `GeoDataFrame`
-    - geopandas `GeoSeries`
-    - numpy array of Shapely objects
-    - Single Shapely object
+    - geopandas `GeoDataFrame`.
+    - geopandas `GeoSeries`.
+    - numpy array of Shapely objects.
+    - Single Shapely object.
+    - A DuckDB query with a spatial column from DuckDB Spatial.
+
+        !!! warning
+
+            The DuckDB query must be run with
+            [`duckdb.sql()`](https://duckdb.org/docs/api/python/reference/#duckdb.sql)
+            or
+            [`duckdb.DuckDBPyConnection.sql()`](https://duckdb.org/docs/api/python/reference/#duckdb.DuckDBPyConnection.sql)
+            and not with `duckdb.execute()` or `duckdb.DuckDBPyConnection.execute()`.
+
+            For example
+
+            ```py
+            import duckdb
+            from lonboard import viz
+
+            sql = "SELECT * FROM spatial_table;"
+            query = duckdb.sql(sql)
+            viz(query)
+            ```
+
+            If you're using a custom connection, ensure you pass in the `con` parameter:
+
+            ```py
+            import duckdb
+            from lonboard import viz
+
+            con = duckdb.connect()
+            sql = "SELECT * FROM spatial_table;"
+            query = con.sql(sql)
+            viz(query, con=con)
+            ```
+
+            You can also render an entire table by using the `table()` method:
+
+            ```py
+            import duckdb
+            from lonboard import viz
+
+            con = duckdb.connect()
+            con.execute("CREATE TABLE spatial_table AS ...;")
+            viz(con.table(), con=con)
+            ```
+
+        !!! warning
+
+            DuckDB Spatial does not currently expose coordinate reference system
+            information, so the user must ensure that data has been reprojected to
+            EPSG:4326.
+
     - Any Python class with a `__geo_interface__` property conforming to the
         [Geo Interface protocol](https://gist.github.com/sgillies/2217756).
     - `dict` holding GeoJSON-like data.
-    - pyarrow `Table` with a geometry column marked with a GeoArrow extension type
-    - pyarrow `Array` marked with a [GeoArrow extension type defined by geoarrow-pyarrow](https://geoarrow.org/geoarrow-python/main/pyarrow.html#geoarrow.pyarrow.GeometryExtensionType)
+    - pyarrow `Table` with a geometry column marked with a
+        [GeoArrow](https://geoarrow.org/) extension type.
+    - pyarrow `Array` marked with a [GeoArrow extension type defined by geoarrow-pyarrow](https://geoarrow.org/geoarrow-python/main/pyarrow.html#geoarrow.pyarrow.GeometryExtensionType).
 
     Alternatively, you can pass a `list` or `tuple` of any of the above inputs.
 
@@ -127,8 +182,12 @@ def viz(
           [`PolygonLayer`][lonboard.PolygonLayer]s.
         map_kwargs: a `dict` of parameters to pass down to the generated
           [`Map`][lonboard.Map].
+        con: the active DuckDB connection. This is necessary in some cases when passing
+            in a DuckDB query. In particular, if you're using a non-global DuckDB
+            connection and if your SQL query outputs the default `GEOMETRY` type.
 
-    For more control over rendering, construct `Map` and `Layer` objects directly.
+    For more control over rendering, construct [`Map`][lonboard.Map] and `Layer` objects
+    directly.
 
     Returns:
         widget visualizing the provided data.
@@ -137,26 +196,26 @@ def viz(
     shuffle(color_ordering)
 
     if isinstance(data, (list, tuple)):
-        layers = [
-            create_layer_from_data_input(
+        layers: List[Union[ScatterplotLayer, PathLayer, PolygonLayer]] = []
+        for i, item in enumerate(data):
+            ls = create_layers_from_data_input(
                 item,
                 _viz_color=color_ordering[i % len(color_ordering)],
                 scatterplot_kwargs=scatterplot_kwargs,
                 path_kwargs=path_kwargs,
                 polygon_kwargs=polygon_kwargs,
+                con=con,
             )
-            for i, item in enumerate(data)
-        ]
+            layers.extend(ls)
     else:
-        layers = [
-            create_layer_from_data_input(
-                data,
-                _viz_color=color_ordering[0],
-                scatterplot_kwargs=scatterplot_kwargs,
-                path_kwargs=path_kwargs,
-                polygon_kwargs=polygon_kwargs,
-            )
-        ]
+        layers = create_layers_from_data_input(
+            data,
+            _viz_color=color_ordering[0],
+            scatterplot_kwargs=scatterplot_kwargs,
+            path_kwargs=path_kwargs,
+            polygon_kwargs=polygon_kwargs,
+            con=con,
+        )
 
     map_kwargs = {} if not map_kwargs else map_kwargs
 
@@ -166,9 +225,34 @@ def viz(
     return Map(layers=layers, **map_kwargs)
 
 
-def create_layer_from_data_input(
-    data: VizDataInput, **kwargs
-) -> Union[ScatterplotLayer, PathLayer, PolygonLayer]:
+DUCKDB_PY_CONN_ERROR = dedent("""\
+    Must pass in DuckDBPyRelation object, not DuckDBPyConnection.
+
+    Instead of using `duckdb.execute()` or `con.execute()`, use `duckdb.sql()` or
+    `con.sql()`.
+
+    If using `con.sql()`, ensure you pass the `con` into the `viz()` function:
+
+    ```
+    viz(con.sql("SELECT * FROM table;", con=con))
+    ```
+
+    Alternatively, you can call the `table()` method of `con`:
+
+    ```
+    viz(con.table("table_name", con=con))
+    ```
+    """)
+
+
+def create_layers_from_data_input(
+    data: VizDataInput, *, con: Optional[duckdb.DuckDBPyConnection] = None, **kwargs
+) -> List[Union[ScatterplotLayer, PathLayer, PolygonLayer]]:
+    """Create one or more renderable layers from data input.
+
+    This helper function can create multiple layers in the case of mixed input.
+    """
+
     # geopandas GeoDataFrame
     if (
         data.__class__.__module__.startswith("geopandas")
@@ -182,6 +266,19 @@ def create_layer_from_data_input(
         and data.__class__.__name__ == "GeoSeries"
     ):
         return _viz_geopandas_geoseries(data, **kwargs)  # type: ignore
+
+    # duckdb DuckDBPyRelation
+    if (
+        data.__class__.__module__.startswith("duckdb")
+        and data.__class__.__name__ == "DuckDBPyRelation"
+    ):
+        return _viz_duckdb_relation(data, con=con, **kwargs)  # type: ignore
+
+    if (
+        data.__class__.__module__.startswith("duckdb")
+        and data.__class__.__name__ == "DuckDBPyConnection"
+    ):
+        raise TypeError(DUCKDB_PY_CONN_ERROR)
 
     # pyarrow table
     if isinstance(data, pa.Table):
@@ -235,40 +332,55 @@ def create_layer_from_data_input(
 
 def _viz_geopandas_geodataframe(
     data: gpd.GeoDataFrame, **kwargs
-) -> Union[ScatterplotLayer, PathLayer, PolygonLayer]:
-    table = geopandas_to_geoarrow(data)
-    return _viz_geoarrow_table(table, **kwargs)
+) -> List[Union[ScatterplotLayer, PathLayer, PolygonLayer]]:
+    layers: List[Union[ScatterplotLayer, PathLayer, PolygonLayer]] = []
+    for partial_gdf in split_mixed_gdf(data):
+        table = geopandas_to_geoarrow(partial_gdf)
+        layers.extend(_viz_geoarrow_table(table, **kwargs))
+
+    return layers
 
 
 def _viz_geopandas_geoseries(
     data: gpd.GeoSeries, **kwargs
-) -> Union[ScatterplotLayer, PathLayer, PolygonLayer]:
+) -> List[Union[ScatterplotLayer, PathLayer, PolygonLayer]]:
     import geopandas as gpd
 
-    gdf = gpd.GeoDataFrame(geometry=data)
-    table = geopandas_to_geoarrow(gdf)
+    gdf = gpd.GeoDataFrame(geometry=data)  # type: ignore
+    return _viz_geopandas_geodataframe(gdf, **kwargs)
+
+
+def _viz_duckdb_relation(
+    data: duckdb.DuckDBPyRelation,
+    con: Optional[duckdb.DuckDBPyConnection] = None,
+    **kwargs,
+) -> List[Union[ScatterplotLayer, PathLayer, PolygonLayer]]:
+    from lonboard._geoarrow._duckdb import from_duckdb
+
+    table = from_duckdb(data, con=con)
     return _viz_geoarrow_table(table, **kwargs)
 
 
 def _viz_shapely_scalar(
     data: shapely.geometry.base.BaseGeometry, **kwargs
-) -> Union[ScatterplotLayer, PathLayer, PolygonLayer]:
+) -> List[Union[ScatterplotLayer, PathLayer, PolygonLayer]]:
     return _viz_shapely_array(np.array([data]), **kwargs)
 
 
 def _viz_shapely_array(
     data: NDArray[np.object_], **kwargs
-) -> Union[ScatterplotLayer, PathLayer, PolygonLayer]:
-    # TODO: pass include_z?
-    field, geom_arr = construct_geometry_array(data)
-    schema = pa.schema([field])
-    table = pa.Table.from_arrays([geom_arr], schema=schema)
-    return _viz_geoarrow_table(table, **kwargs)
+) -> List[Union[ScatterplotLayer, PathLayer, PolygonLayer]]:
+    # Note: for now we pass this through a GeoDataFrame to handle mixed-type geometry
+    # arrays. Longer term we should do this without a GeoPandas dependency.
+    import geopandas as gpd
+
+    gdf = gpd.GeoDataFrame(geometry=data)  # type: ignore
+    return _viz_geopandas_geodataframe(gdf, **kwargs)
 
 
 def _viz_geo_interface(
     data: dict, **kwargs
-) -> Union[ScatterplotLayer, PathLayer, PolygonLayer]:
+) -> List[Union[ScatterplotLayer, PathLayer, PolygonLayer]]:
     if data["type"] in [
         "Point",
         "LineString",
@@ -287,6 +399,11 @@ def _viz_geo_interface(
         return _viz_geoarrow_table(table.append_column(field, geom_arr), **kwargs)
 
     if data["type"] == "FeatureCollection":
+        # We currently take a FeatureCollection through GeoPandas so that we can handle
+        # mixed-geometry type collections
+        import geopandas as gpd
+        import pandas as pd
+
         attribute_columns_struct = pa.array(
             [feature["properties"] for feature in data["features"]]
         )
@@ -295,15 +412,16 @@ def _viz_geo_interface(
         arrays = []
         for field_idx in range(attribute_columns_struct.type.num_fields):
             fields.append(attribute_columns_struct.type.field(field_idx))
-            arrays.append(pc.struct_field(attribute_columns_struct, field_idx))
+            arrays.append(pc.struct_field(attribute_columns_struct, field_idx))  # type: ignore
 
         table = pa.Table.from_arrays(arrays, schema=pa.schema(fields))
+        df = table.to_pandas(types_mapper=pd.ArrowDtype)
 
         shapely_geom_arr = shapely.from_geojson(
             [json.dumps(feature["geometry"]) for feature in data["features"]]
         )
-        field, geom_arr = construct_geometry_array(shapely_geom_arr)
-        return _viz_geoarrow_table(table.append_column(field, geom_arr), **kwargs)
+        gdf = gpd.GeoDataFrame(df, geometry=shapely_geom_arr)  # type: ignore
+        return _viz_geopandas_geodataframe(gdf, **kwargs)
 
     geo_interface_type = data["type"]
     raise ValueError(f"type '{geo_interface_type}' not supported.")
@@ -312,7 +430,7 @@ def _viz_geo_interface(
 def _viz_geoarrow_array(
     data: ArrowArrayExportable,
     **kwargs,
-) -> Union[ScatterplotLayer, PathLayer, PolygonLayer]:
+) -> List[Union[ScatterplotLayer, PathLayer, PolygonLayer]]:
     schema_capsule, array_capsule = data.__arrow_c_array__()
 
     # If the user doesn't have pyarrow extension types registered for geoarrow types,
@@ -362,9 +480,25 @@ def _viz_geoarrow_table(
     scatterplot_kwargs: Optional[ScatterplotLayerKwargs] = None,
     path_kwargs: Optional[PathLayerKwargs] = None,
     polygon_kwargs: Optional[PolygonLayerKwargs] = None,
-) -> Union[ScatterplotLayer, PathLayer, PolygonLayer]:
+) -> List[Union[ScatterplotLayer, PathLayer, PolygonLayer]]:
     table = remove_extension_classes(table)
-    table = parse_wkb_table(table)
+    parsed_tables = parse_wkb_table(table)
+    if len(parsed_tables) > 1:
+        output: List[Union[ScatterplotLayer, PathLayer, PolygonLayer]] = []
+        for parsed_table in parsed_tables:
+            output.extend(
+                _viz_geoarrow_table(
+                    parsed_table,
+                    _viz_color=_viz_color,
+                    scatterplot_kwargs=scatterplot_kwargs,
+                    path_kwargs=path_kwargs,
+                    polygon_kwargs=polygon_kwargs,
+                )
+            )
+
+        return output
+    else:
+        table = parsed_tables[0]
 
     geometry_column_index = get_geometry_column_index(table.schema)
     geometry_field = table.schema.field(geometry_column_index)
@@ -394,7 +528,7 @@ def _viz_geoarrow_table(
             elif len(table) <= 1_000_000:
                 scatterplot_kwargs["opacity"] = 0.5
 
-        return ScatterplotLayer(table=table, **scatterplot_kwargs)
+        return [ScatterplotLayer(table=table, **scatterplot_kwargs)]
 
     elif geometry_ext_type in [
         EXTENSION_NAME.LINESTRING,
@@ -423,7 +557,7 @@ def _viz_geoarrow_table(
             elif len(table) <= 100_000:
                 path_kwargs["opacity"] = 0.5
 
-        return PathLayer(table=table, **path_kwargs)
+        return [PathLayer(table=table, **path_kwargs)]
 
     elif geometry_ext_type in [EXTENSION_NAME.POLYGON, EXTENSION_NAME.MULTIPOLYGON]:
         polygon_kwargs = {} if not polygon_kwargs else polygon_kwargs
@@ -451,6 +585,6 @@ def _viz_geoarrow_table(
             else:
                 polygon_kwargs["line_width_min_pixels"] = 0.2
 
-        return PolygonLayer(table=table, **polygon_kwargs)
+        return [PolygonLayer(table=table, **polygon_kwargs)]
 
     raise ValueError(f"Unsupported extension type: '{geometry_ext_type}'.")